+++ src/main/java/com/jcloisterzone/wsio/message/SetExpansionMessage.java
@@ -8,14 +8,14 @@ public class SetExpansionMessage implements WsInGameMessage {
 
     private String gameId;
     private Expansion expansion;
-    private boolean enabled;
+    private int count;
 
     public SetExpansionMessage() {
     }
 
-    public SetExpansionMessage(Expansion expansion, boolean enabled) {
+    public SetExpansionMessage(Expansion expansion, int count) {
         this.expansion = expansion;
-        this.enabled = enabled;
+        this.count = count;
     }
 
     @Override
@@ -36,12 +36,11 @@ public class SetExpansionMessage implements WsInGameMessage {
         this.expansion = expansion;
     }
 
-    public boolean isEnabled() {
-        return enabled;
+    public int getCount() {
+        return count;
     }
 
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
+    public void setCount(int count) {
+        this.count = count;
     }
-
 }